+++ src/pages/report/Overview/OverviewDate/index.tsx
@@ -3,7 +3,6 @@ import { TabContext, TabList, TabPanel } from '@mui/lab';
 // material
 import { Box, Card, Tab } from '@mui/material';
 import useLocales from 'hooks/useLocales';
-import moment from 'moment';
 import { useSnackbar } from 'notistack';
 import ReportBtn from 'pages/report/components/ReportBtn';
 import ReportDatePicker from 'pages/report/components/ReportDatePicker';
@@ -11,9 +10,10 @@ import ReportPage from 'pages/report/components/ReportPage';
 import React, { useState } from 'react';
 // components
 import { useNavigate } from 'react-router-dom';
-import RevenueOverview from '../components/RevenueOverview';
+import { fDate, fTime } from 'utils/formatTime';
 import EmployeeStatistics from '../components/EmployeeStatistics';
 import ProductSaleDetail from '../components/ProductSaleDetail';
+import RevenueOverview from '../components/RevenueOverview';
 import TopStoreRevenue from '../components/TopStoreRevenue';
 
 // icons
@@ -30,16 +30,8 @@ export default function OverviewDate() {
 
   return (
     <ReportPage
-      title={`Tổng quan ngày: ${date.toLocaleDateString('vi-VI', {
-        year: 'numeric',
-        month: '2-digit',
-        day: '2-digit'
-      })}`}
-      content={
-        date.toDateString() === today.toDateString()
-          ? `Tính đến: ${moment().format('hh:mm:ss')}`
-          : ''
-      }
+      title={`Tổng quan ngày: ${fDate(date)}`}
+      content={date.toDateString() === today.toDateString() ? `Tính đến: ${fTime(date)}` : ``}
       actions={[
         <ReportDatePicker
           key="choose-day"