+++ src/utils/formatTime.ts
@@ -19,3 +19,7 @@ export function fToNow(date: string | number | Date) {
     addSuffix: true
   });
 }
+
+export function fTime(date: string | number | Date) {
+  return format(new Date(date), 'hh:mm:ss');
+}