+++ changeLog.md
@@ -1,8 +1,14 @@
+#### `2022-06-25` `Phuoclt:`
+###### Report system
+- 🛠 Refactor code: using the same format function and clear code 
+- 🌐 Remove a format currency vnd `fCurrencyVN`
+- 🌐 Add a new format time only `fTime`
 #### `2022-06-24` `Phuoclt:`
 ###### Report system
 - 🆕 Add `productSale.ts` and `productProgress` in type to manage type of productSale and productProgress
 - 🛠 Refactor code: create `config.ts` in productSale and productProgress component to manage column Resotable 
 - 💄 Change the same spacing product and overview component and add tab ui in productSale and productProgress
+- 🌐 Create a format currency vnd `fCurrencyVN`
 
 #### `2022-06-22` `Phuoclt:`
 ###### Report system